Avoid throwing the same exception instances in parallel
